+++ replay_trajectory_classification/likelihoods/spiking_likelihood_glm.py
@@ -105,7 +105,7 @@ def fit_glm(
         penalty = np.ones((design_matrix.shape[1],)) * penalty
         penalty[0] = 0.0  # don't penalize the intercept
     else:
-        penalty = np.finfo(np.float).eps
+        penalty = np.finfo(float).eps
     return penalized_IRLS(
         design_matrix,
         response.squeeze(),
@@ -188,7 +188,7 @@ def estimate_spiking_likelihood(
 
     log_likelihood = combined_likelihood(spikes, conditional_intensity)
 
-    mask = np.ones_like(is_track_interior, dtype=np.float)
+    mask = np.ones_like(is_track_interior, dtype=float)
     mask[~is_track_interior] = np.nan
 
     return log_likelihood * mask
